Moved expr tests to expr/mod.rs
